Added a zshrc and vimrc customizer

+++ setup.sh
@@ -1,10 +1,17 @@
 #!/bin/bash
 
-# Clone dotfiles repo
-git clone https://github.com/grant-wade/dotfiles.git .dotfiles
+# Print where the install will be happening
+printf "Installing files to %s\n" $HOME
 
-# Run setup script
-python2 ~/.dotfiles/install.py
+# See if user wants to continue
+read -p "Continue? [yn]: " confirm
+if [[ $confirm != "y" ]]; then
+    echo "Exiting"
+    exit 0
+fi
+
+# Clone dotfiles repo
+git clone https://github.com/grant-wade/dotfiles.git $HOME/.dotfiles
 
 # Download miniconda install script
 curl -O https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
@@ -14,6 +21,12 @@ bash Miniconda3-latest-Linux-x86_64.sh
 
 # Remove miniconda installer
 rm Miniconda3-latest-Linux-x86_64.sh
+
+# Run setup script
+$HOME/.miniconda3/bin/python $HOME/.dotfiles/install.py
+
+# Customize dotfiles
+$HOME/.miniconda3/bin/python $HOME/.dotfiles/file_creator.py
 
 # Link all the needed files
 ln -sf ~/.dotfiles/zshrc ~/.zshrc
